requirements.txt
anyio==3.6.1
fastapi==0.85.0
greenlet==1.1.2
grpcio==1.47.0
grpcio-tools==1.47.0
idna==3.4
protobuf==3.20.1
psycopg2-binary==2.9.3
pydantic==1.8.2
python-dotenv==0.20.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.40
starlette==0.20.4
fastapi==0.85.0
greenlet==1.1.2
grpcio==1.47.0
grpcio-tools==1.47.0
idna==3.4
protobuf==3.20.1
psycopg2-binary==2.9.3
pydantic==1.8.2
python-dotenv==0.20.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.40
starlette==0.20.4
autopep8==1.7.0
pycodestyle==2.9.1
SQLAlchemy-Utils==0.37.8
toml==0.10.2
XlsxWriter==3.0.3
uvicorn~=0.18.3